requirements.txt
streamlit
os
dotenv
google-genai
